requirements.txt
numpy
scipy
matplotlib
Flask
Pillow
pdfkit
opencv-python
PyPDF2
